Move Forms types, handlers and providers to the right place
