+++ src/PrestaShopBundle/Form/Admin/Configure/ShopParameters/General/MaintenanceFormDataProvider.php
@@ -23,7 +23,7 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  * International Registered Trademark & Property of PrestaShop SA
  */
-namespace PrestaShopBundle\Form\Admin\ShopParameters\General;
+namespace PrestaShopBundle\Form\Admin\Configure\ShopParameters\General;
 
 use PrestaShop\PrestaShop\Adapter\Shop\MaintenanceConfiguration;
 use PrestaShop\PrestaShop\Core\Form\FormDataProviderInterface;
