+++ src/PrestaShopBundle/Form/Admin/Configure/ShopParameters/ProductPreferences/GeneralType.php
@@ -1,5 +1,5 @@
 <?php
-/**
+/*
  * 2007-2018 PrestaShop
  *
  * NOTICE OF LICENSE
@@ -7,7 +7,7 @@
  * This source file is subject to the Open Software License (OSL 3.0)
  * that is bundled with this package in the file LICENSE.txt.
  * It is also available through the world-wide-web at this URL:
- * https://opensource.org/licenses/OSL-3.0
+ * http://opensource.org/licenses/osl-3.0.php
  * If you did not receive a copy of the license and are unable to
  * obtain it through the world-wide-web, please send an email
  * to license@prestashop.com so we can send you a copy immediately.
@@ -18,37 +18,52 @@
  * versions in the future. If you wish to customize PrestaShop for your
  * needs please refer to http://www.prestashop.com for more information.
  *
- * @author    PrestaShop SA <contact@prestashop.com>
- * @copyright 2007-2018 PrestaShop SA
- * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
- * International Registered Trademark & Property of PrestaShop SA
+ *  @author PrestaShop SA <contact@prestashop.com>
+ *  @copyright  2007-2018 PrestaShop SA
+ *  @license    http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ *  International Registered Trademark & Property of PrestaShop SA
  */
 
-namespace PrestaShopBundle\Form\Admin\ShopParameters\CustomerPreferences;
+namespace PrestaShopBundle\Form\Admin\Configure\ShopParameters\ProductPreferences;
 
 use PrestaShopBundle\Form\Admin\Type\SwitchType;
 use PrestaShopBundle\Form\Admin\Type\TextWithUnitType;
 use PrestaShopBundle\Form\Admin\Type\TranslatorAwareType;
+use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+use Symfony\Component\Form\Extension\Core\Type\IntegerType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
  * Class generates "General" form
- * in "Configure > Shop Parameters > Customer Settings" page.
+ * in "Configure > Shop Parameters > Product Settings" page.
  */
 class GeneralType extends TranslatorAwareType
 {
+    /**
+     * {@inheritdoc}
+     */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('redisplay_cart_at_login', SwitchType::class)
-            ->add('send_email_after_registration', SwitchType::class)
-            ->add('password_reset_delay', TextWithUnitType::class, [
-                'unit' => $this->trans('minutes', 'Admin.Shopparameters.Feature'),
+            ->add('catalog_mode', SwitchType::class)
+            ->add('new_days_number', IntegerType::class, [
+                'required' => false,
             ])
-            ->add('enable_b2b_mode', SwitchType::class)
-            ->add('ask_for_birthday', SwitchType::class)
-            ->add('enable_offers', SwitchType::class)
+            ->add('short_description_limit', TextWithUnitType::class, [
+                'required' => false,
+                'unit' => $this->trans('characters', 'Admin.Shopparameters.Help'),
+            ])
+            ->add('quantity_discount', ChoiceType::class, [
+                'choices' => [
+                    'Products' => 0,
+                    'Combinations' => 1,
+                ],
+                'choice_translation_domain' => 'Admin.Global',
+                'required' => true,
+            ])
+            ->add('force_friendly_url', SwitchType::class)
+            ->add('default_status', SwitchType::class)
         ;
     }
 
@@ -67,6 +82,6 @@ class GeneralType extends TranslatorAwareType
      */
     public function getBlockPrefix()
     {
-        return 'customer_preferences_general_block';
+        return 'product_preferences_general_block';
     }
 }
