+++ src/PrestaShopBundle/Resources/config/services/form/form_data_provider.yml
@@ -3,13 +3,13 @@ services:
         public: true
 
     prestashop.admin.customer_preferences.data_provider:
-        class: 'PrestaShopBundle\Form\Admin\ShopParameters\CustomerPreferences\CustomerPreferencesDataProvider'
+        class: 'PrestaShopBundle\Form\Admin\Configure\ShopParameters\CustomerPreferences\CustomerPreferencesDataProvider'
         arguments:
             - '@prestashop.adapter.customer.customer_configuration'
             - '@translator'
 
     prestashop.admin.product_preferences.data_provider:
-        class: 'PrestaShopBundle\Form\Admin\ShopParameters\ProductPreferences\ProductPreferencesFormDataProvider'
+        class: 'PrestaShopBundle\Form\Admin\Configure\ShopParameters\ProductPreferences\ProductPreferencesFormDataProvider'
         arguments:
             - '@prestashop.adapter.product_general.configuration'
             - '@prestashop.adapter.product_pagination.configuration'
@@ -27,7 +27,7 @@ services:
             - '@prestashop.adapter.caching.configuration'
 
     prestashop.adapter.maintenance.form_provider:
-        class: 'PrestaShopBundle\Form\Admin\ShopParameters\General\MaintenanceFormDataProvider'
+        class: 'PrestaShopBundle\Form\Admin\Configure\ShopParameters\General\MaintenanceFormDataProvider'
         arguments:
             - '@prestashop.adapter.maintenance.configuration'
 
@@ -39,7 +39,7 @@ services:
             - '@prestashop.adapter.notifications.configuration'
 
     prestashop.adapter.preferences.form_provider:
-        class: 'PrestaShopBundle\Form\Admin\ShopParameters\General\PreferencesFormDataProvider'
+        class: 'PrestaShopBundle\Form\Admin\Configure\ShopParameters\General\PreferencesFormDataProvider'
         arguments:
             - '@prestashop.adapter.preferences.configuration'
